random32: rename random32 to prandom
commit 496f2f93b1cc286f5a4f4f9acdc1e5314978683f upstream.

This renames all random32 functions to have 'prandom_' prefix as follows:

    void prandom_seed(u32 seed);  /* rename from srandom32() */
    u32 prandom_u32(void);                /* rename from random32() */
    void prandom_seed_state(struct rnd_state *state, u64 seed);
                                    /* rename from prandom32_seed() */
    u32 prandom_u32_state(struct rnd_state *state);
                                    /* rename from prandom32() */

The purpose of this renaming is to prevent some kernel developers from
assuming that prandom32() and random32() might imply that only
prandom32() was the one using a pseudo-random number generator by
prandom32's "p", and the result may be a very embarassing security
exposure.  This concern was expressed by Theodore Ts'o.